+++ src/main/java/jewellery/inventory/dto/response/resource/ResourceResponseDto.java
@@ -2,6 +2,7 @@ package jewellery.inventory.dto.response.resource;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import java.math.BigDecimal;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -23,6 +24,6 @@ public class ResourceResponseDto {
     private UUID id;
     private String clazz;
     private String quantityType;
-    private double pricePerQuantity;
+    private BigDecimal pricePerQuantity;
     private String note;
 }
